services/dashboard/requirements.txt
# Flask Dashboard Requirements
# Python 3.11+

# Web Framework
Flask==3.0.0
Flask-Cors==4.0.0

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Environment Variables
python-dotenv==1.0.0

# Template Rendering
Jinja2==3.1.2

# Development Tools
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
